seed builds docker before testing against custom fixture
